SMART FASHIN RECOMENDER APPLICATION
README.md
@ -1,8 +1,32 @@
# IBM-Project-49293-1660817543
Smart Fashion Recommender Application

Team :
    Sasisharanavathy.M - TL
    Bagavathiathal.P - TM
    Bhavadharani.K - TM
    Deepak.S -TM
# Smart Fashion Recommender Application

&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;We have come up with a new innovative solution through which you can directly do your online shopping based on your choice without any search. It can be done by using the chatbot.

## Team Members:
#### 1. Sasisharanavathy.M - TL
#### 2. Bagavathiathal.P - TM
#### 3. Bhavadharani.K - TM
#### 4. Deepak.S -TM

## Abstract :

&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Nowadays, fashion applications and e-commerce are growing more and more.And it also has some problems when finding the customer's wanted product in the web applications. Having a chatbot that understands the algorithm of a specific application
can be of great aid. We are implementing such a chat bot in a web application, which is
fed with the knowledge of the application’s algorithm and helps the user completely
from finding their needs to processing the payment and initiating delivery. It works as an advanced filter search that can bring the user what they want with the help of pictorial and named representation by getting simple user information and activities. The
application also has two main UI interactions: one is the user panel and the other one is
the admin panel. Users can interact with the chat bot to search for products, order them
from the manufacturer or distributor through chatbot AI, and it can also make payment
transactions, track the delivery, and so on. The admin interface enables the user to
upload products' details ,user details, orders and find how many products have been
bought; supervise the stock availability; and interact with the buyer regarding the
product reviews.


## Features of Chatbot :
>- Using chatbot we can manage user's choices and orders.
>- The chatbot can give recommendations to the users based on their interests.
>- It can promote the best deals and offers on that day.
>- It will store the customer's details and orders in the database.
>- The chatbot will send a notification to customers if the order is confirmed.
>- Chatbots can also help in collecting customer feedback.